Require same major version of neutron
Release cycles for the -aas project are in sync with neutron, as well as
major version numbers. Since the -aas code directly imports from neutron
module, we should make sure that the major versions are identical,
otherwise we can get into a situation where e.g. vpnaas code is a lot
older than neutron and is not compatible with it, breaking the plugin,
as well as other tools that happen to import vpnaas code during their
operations (f.e. neutron-db-manage does it).

Sadly, vpnaas is not part of neutron stadium, and doesn't guarantee a
release per neutron release; it also hasn't tagged anything for Pike
just yet. So instead of deducing neutron major version from the package
itself, we need to hardcode it once per release.

Change-Id: I01e97ae3a6db40fefdbb85f0e9b4a1a9d3c39600