Update Conflicts/Requires on openstack-neutron
Neutron 14.0.0 beta is released in [1], but neutron-vpnaas
is not released, this leads neutron-vpnaas 13 to conflict with
neutron 14. neutron-vpnaas is now used to be released around
openstack-neutron so there are rare changes they will differ
in major release for long and hence break compatibility.

Also use %{epoch} instead of hardcoding neutron epoch to 1.

[1] https://review.openstack.org/#/c/613444/

Change-Id: I2d3666014e9da57be754ae1fad11e58a3d613a46